third_party/blink/web_tests/external/wpt/css/css-text/parsing/white-space-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['white-space'] = "balance" should not set the property value
  assert_equals: expected "" but got "balance"
[FAIL] e.style['white-space'] = "collapse balance" should not set the property value
  assert_equals: expected "" but got "balance"
[FAIL] e.style['white-space'] = "balance collapse" should not set the property value
  assert_equals: expected "" but got "balance"
[FAIL] e.style['white-space'] = "preserve balance" should not set the property value
  assert_equals: expected "" but got "preserve balance"
[FAIL] e.style['white-space'] = "balance preserve" should not set the property value
  assert_equals: expected "" but got "preserve balance"
Harness: the test ran to completion.
